+++ src/navigation/admin/authorization.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useMemo } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
-import AuthorizationPage from '../../components/Admin/AuthorizationPage';
+import AuthorizationPage from '../../components/Admin/Authorization/AuthorizationPage';
 import { FourOuFour, PageProps } from '../../pages';
 
 interface AuthorizationRouteProps extends PageProps {}
